requirements.txt
flask
torch
keras
numpy
pillow
gunicorn